requirements.txt
python-dotenv
fastapi[all]
langchain
langchain-ollama
langchain-community
langchain-deepseek
ultralytics
redis
PyJWT